Add cassandra unit patch for whole integration tests suite
