+++ src/cassandra-open-service-broker/.mvn/org/cassandraunit/cassandra-unit-parent/3.3.0.3-SNAPSHOT/maven-metadata-local.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<metadata modelVersion="1.1.0">
+  <groupId>org.cassandraunit</groupId>
+  <artifactId>cassandra-unit-parent</artifactId>
+  <version>3.3.0.3-SNAPSHOT</version>
+  <versioning>
+    <snapshot>
+      <localCopy>true</localCopy>
+    </snapshot>
+    <lastUpdated>20171103153347</lastUpdated>
+    <snapshotVersions>
+      <snapshotVersion>
+        <extension>pom</extension>
+        <value>3.3.0.3-SNAPSHOT</value>
+        <updated>20171103153347</updated>
+      </snapshotVersion>
+    </snapshotVersions>
+  </versioning>
+</metadata>
