+++ src/cassandra-open-service-broker/.mvn/org/cassandraunit/cassandra-unit-spring/3.3.0.3-SNAPSHOT/cassandra-unit-spring-3.3.0.3-SNAPSHOT.pom.lastUpdated
@@ -1,0 +1,6 @@
+#NOTE: This is an Aether internal implementation file, its format can be changed without prior notice.
+#Fri Nov 03 16:33:07 CET 2017
+https\://repo.spring.io/libs-snapshot/.error=
+http\://dvedv907.rouen.francetelecom.fr\:8081/nexus/content/groups/public/.lastUpdated=1509723187285
+https\://repo.spring.io/libs-snapshot/.lastUpdated=1509723187892
+http\://dvedv907.rouen.francetelecom.fr\:8081/nexus/content/groups/public/.error=
