+++ src/cassandra-open-service-broker/.mvn/org/cassandraunit/cassandra-unit/3.3.0.3-SNAPSHOT/maven-metadata-local.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<metadata modelVersion="1.1.0">
+  <groupId>org.cassandraunit</groupId>
+  <artifactId>cassandra-unit</artifactId>
+  <version>3.3.0.3-SNAPSHOT</version>
+  <versioning>
+    <snapshot>
+      <localCopy>true</localCopy>
+    </snapshot>
+    <lastUpdated>20171103153409</lastUpdated>
+    <snapshotVersions>
+      <snapshotVersion>
+        <classifier>bin</classifier>
+        <extension>tar.gz</extension>
+        <value>3.3.0.3-SNAPSHOT</value>
+        <updated>20171103153409</updated>
+      </snapshotVersion>
+      <snapshotVersion>
+        <extension>jar</extension>
+        <value>3.3.0.3-SNAPSHOT</value>
+        <updated>20171103153409</updated>
+      </snapshotVersion>
+      <snapshotVersion>
+        <extension>pom</extension>
+        <value>3.3.0.3-SNAPSHOT</value>
+        <updated>20171103153409</updated>
+      </snapshotVersion>
+    </snapshotVersions>
+  </versioning>
+</metadata>
