+++ src/cassandra-open-service-broker/.mvn/org/cassandraunit/cassandra-unit/3.3.0.3-SNAPSHOT/resolver-status.properties
@@ -1,0 +1,6 @@
+#NOTE: This is an Aether internal implementation file, its format can be changed without prior notice.
+#Fri Nov 03 16:33:06 CET 2017
+maven-metadata-paas.public.xml.lastUpdated=1509723185850
+maven-metadata-spring-libs-snapshot.xml.lastUpdated=1509723186216
+maven-metadata-spring-libs-snapshot.xml.error=
+maven-metadata-paas.public.xml.error=
