+++ src/cassandra-open-service-broker/pom.xml
@@ -43,7 +43,8 @@
         <dependency>
           <groupId>org.springframework.data</groupId>
           <artifactId>spring-data-cassandra</artifactId>
-          <version>1.5.6.RELEASE</version>
+          <version>1.5.7.RELEASE</version>
+          <!--<version>2.0.0.RC3</version>-->
         </dependency>
 		
         <dependency>
@@ -74,14 +75,14 @@
         <dependency>
             <groupId>org.cassandraunit</groupId>
             <artifactId>cassandra-unit</artifactId>
-            <version>3.1.3.2</version>
+            <version>3.3.0.3-SNAPSHOT</version>
             <scope>test</scope>
         </dependency>
 
         <dependency>
             <groupId>org.cassandraunit</groupId>
             <artifactId>cassandra-unit-spring</artifactId>
-            <version>3.1.3.2</version>
+            <version>3.3.0.3-SNAPSHOT</version>
             <scope>test</scope>
         </dependency>
 
@@ -96,5 +97,11 @@
 		</plugins>
 	</build>
 
+    <repositories>
+        <repository>
+            <id>my-local-repo</id>
+            <url>file://${basedir}/.mvn</url>
+        </repository>
+    </repositories>
 
 </project>
