+++ src/cassandra-open-service-broker/src/test/resources/application-test_embedded.properties
@@ -1,5 +1,5 @@
 spring.data.cassandra.contact-points=127.0.0.1
-spring.data.cassandra.port=9042
+spring.data.cassandra.port=9142
 spring.data.cassandra.username=cassandra
 spring.data.cassandra.password=cassandra
 cassandra.startupTimeoutInSeconds=50
