--- src/cassandra-open-service-broker/src/test/resources/cassandra-topology.properties
@@ -1,4 +1,0 @@
-#properties file used with PropertyFileSnitch to give topology to cassandra cluster
-#this way we can test distributed cluster giving it ip = data center:rack notation for
-#each node
-127.0.0.1=DC1:RAC1
