--- src/cassandra-open-service-broker/src/test/resources/readme.txt
@@ -1,8 +1,0 @@
-https://www.tomaz.me/2013/09/21/speeding-up-tests-which-talk-to-cassandra-database.html
-Usage d'un disque mémoire :
-sudo -i
-hostname yd-ijly7474
-mkdir /tmp/ramdisk
-chmod 777 /tmp/ramdisk
-sudo mount -t tmpfs -o size=384M tmpfs /tmp/ramdisk/
-
